requirements.txt
# 核心依赖
SpeechRecognition==3.10.0    # 降级到稳定版本
pyttsx3>=2.90              # 语音合成
openai==1.3.0              # OpenAI API
python-dotenv>=0.19.0      # 环境变量管理

# 音频处理
PyAudio>=0.2.11            # 音频录制
numpy>=1.21.0              # 数组处理
sounddevice>=0.4.4         # 音频设备管理

# 网络相关
requests>=2.26.0           # HTTP请求
aiohttp>=3.8.1             # 异步HTTP
urllib3>=1.26.7            # HTTP客户端

# 性能监控
psutil>=5.8.0              # 系统和进程监控
memory-profiler>=0.58.0    # 内存分析

# 缓存管理
cachetools>=4.2.4          # 缓存工具
redis>=4.0.2               # Redis客户端（可选，用于分布式缓存）

# 日志和监控
prometheus-client>=0.12.0   # 指标收集（可选）

# 测试相关
pytest>=6.2.5              # 单元测试
pytest-cov==2.12.1         # 测试覆盖率（固定版本以避免依赖冲突）
coverage>=5.2.1,<6.0       # 明确指定 coverage 版本范围
pytest-asyncio>=0.16.0     # 异步测试

# 开发工具
black>=21.12b0             # 代码格式化
flake8>=4.0.1              # 代码检查
mypy>=0.910                # 类型检查

# 文档生成
Sphinx>=4.3.2              # 文档生成（可选）
sphinx-rtd-theme>=1.0.0    # 文档主题（可选）

# 其他工具
tqdm>=4.62.3               # 进度条
colorama>=0.4.4            # 终端着色

# 其他依赖